Criar permissões para `Equipamentos\Cadastro\Marca` Fixes #272

+++ app/Http/Controllers/Admin/Equipamentos/Cadastro/MarcaController.php
@@ -6,6 +6,7 @@ use App\Enums\Equipamentos\Cadastro\StatusCadastro;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Equipamentos\Cadastro\MarcaRequest;
 use App\Models\Equipamentos\Cadastro\Marca;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -13,6 +14,7 @@ class MarcaController extends Controller
 {
     public function inicio()
     {
+        Gate::authorize('ver', Marca::class);
         $marcas = Marca::paginate(10);
         $statusCadastro = StatusCadastro::toArray();
 
@@ -21,12 +23,14 @@ class MarcaController extends Controller
 
     public function criar()
     {
+        Gate::authorize('criar', Marca::class);
         $statusCadastro = StatusCadastro::toArray();
         return Inertia::render('Admin/Equipamentos/Cadastro/Marca/Criar', compact('statusCadastro'));
     }
 
     public function salvar(MarcaRequest $request)
     {
+        Gate::authorize('criar', Marca::class);
         Marca::create($request->all());
 
         return redirect()->route('admin.marcas');
@@ -34,6 +38,7 @@ class MarcaController extends Controller
 
     public function editar(int $id)
     {
+        Gate::authorize('editar', Marca::class);
         $marca = Marca::findOrFail($id);
         $statusCadastro = StatusCadastro::toArray();
 
@@ -42,6 +47,7 @@ class MarcaController extends Controller
 
     public function atualizar(MarcaRequest $request, int $id)
     {
+        Gate::authorize('editar', Marca::class);
         Marca::findOrFail($id)->update($request->all());
 
         return redirect()->route('admin.marcas');
@@ -49,6 +55,7 @@ class MarcaController extends Controller
 
     public function excluir(int $id)
     {
+        Gate::authorize('excluir', Marca::class);
         Marca::findOrFail($id)->delete();
 
         return redirect()->route('admin.marcas');
@@ -56,6 +63,7 @@ class MarcaController extends Controller
 
     public function pesquisar(Request $request)
     {
+        Gate::authorize('ver', Marca::class);
         $marcas = Marca::select('id', 'nome as texto')
             ->whereFullText('nome', $request->input('termo'))
             ->orWhere('nome', 'like', '%' . $request->input('termo') . '%')
@@ -67,6 +75,7 @@ class MarcaController extends Controller
 
     public function salvarAjax(MarcaRequest $request)
     {
+        Gate::authorize('criar', Marca::class);
         $marca = Marca::create($request->all());
 
         return response()->json($marca);
